Changed Main tx_task and sensorNode
main-> removed logic handling for type of sensor as well as includes for individual sensors
sensorNode->changed to wrap all sensors in one singular sensorNode class(for ease of use), also sends data in json. Need to fix ID determination strategy in order to include duplex data transmission, as it is now, root is purely listening, and children are purely sending.